web/requirements.txt
Django==3.1
gunicorn==20.0.4
psycopg2>=2.7,<3.0
gevent
djangorestframework
pylint
pylint-django
